лекция 3
