+++ index.js
@@ -1,26 +1,16 @@
-const fs = require('fs');
-const readline = require('readline');
+const express = require("express");
 
+const app = express();
 
-const readStream = fs.createReadStream('./access.log', 'utf-8');
-const writeStream1 = fs.createWriteStream('./89.123.1.41_request.log');
-const writeStream2 = fs.createWriteStream('./34.48.240.111_request.log');
-
-let numStr = 0;
-
-const rl = readline.createInterface({
-  input:readStream,
-  terminal: true
-});
-
-rl.on('line', (line) => {
-  if(line.includes("89.123.1.41")) {
-    writeStream1.write(line + "\n")
-  }
-
-  if(line.includes("34.48.240.111")) {
-    writeStream2.write(line + "\n")
-  }
-
-  console.log(++numStr)
+app.get('/hello', (req, res)=> {
+  res.json({'name': 'ivan'});
 })
+
+app.get('/html', (req, res)=> {
+  res.sendFile('./index.html');
+})
+
+app.get('/', (req, res)=> {
+  res.sendStatus(200);
+})
+app.listen(3000, () =>  console.log('Server been started http://localhost:3000'))
