--- nodemon.json
@@ -1,6 +1,0 @@
-{
-  "watch": ["src"],
-  "ext": "ts, js",
-  "ignore": ["src/**/*.test.ts"],
-  "exec": "ts-node index.ts"
-}
