+++ package.json
@@ -21,6 +21,7 @@
   "dependencies": {
     "8": "^0.0.1",
     "colors": "^1.4.0",
+    "express": "^4.18.1",
     "inquirer": "8.2",
     "nodemon": "^2.0.19",
     "readline": "^1.3.0",
